Search user css works correctly

+++ resources/views/components/user/user/data.blade.php
@@ -83,7 +83,7 @@
         </div>
     @endif
 
-    @if (Auth::user()->slug === $user->slug)
+    @if (Auth::check() && Auth::user()->slug === $user->slug)
         <div class="actions flex justify-end mt-8">
             <div class="flex justify-end">
                 <a href="#update" class="update-button btn btn-icon btn-one p-2">
@@ -97,7 +97,7 @@
                 </a>
             </div>
     @endif
-    @if (Auth::user()->slug !== $user->slug && isset($user->isFriend) && $user->isFriend === 0 && Auth::user()->id_role === 0)
+    @if (Auth::check() && Auth::user()->slug !== $user->slug && isset($user->isFriend) && $user->isFriend === 0 && Auth::user()->id_role === 0)
         <div class="actions flex justify-end mt-8 mt-4">
             <div class="flex justify-end">
                 <a href="/users/{{ $user->slug }}/friendship/request" class="btn btn-outline btn-one py-2 px-4 ml-4">
@@ -105,7 +105,7 @@
                 </a>
             </div>
     @endif
-    @if (Auth::user()->slug !== $user->slug && isset($user->isFriend) && $user->isFriend === 1 && Auth::user()->id_role === 0)
+    @if (Auth::check() && Auth::user()->slug !== $user->slug && isset($user->isFriend) && $user->isFriend === 1 && Auth::user()->id_role === 0)
         <div class="actions flex justify-end mt-8 mt-4">
             <div class="flex justify-end">
                 @if (Auth::user()->id_user === $user->id_user_request)
@@ -126,7 +126,7 @@
                 @endif
             </div>
     @endif
-    @if (Auth::user()->slug !== $user->slug && isset($user->isFriend) && $user->isFriend === 2 && Auth::user()->id_role === 0)
+    @if (Auth::check() && Auth::user()->slug !== $user->slug && isset($user->isFriend) && $user->isFriend === 2 && Auth::user()->id_role === 0)
         <div class="actions flex justify-end mt-8">
             <div class="flex justify-end">
                 <a href="/users/{{ $user->slug }}/friendship/delete" class="btn btn-outline btn-three py-2 px-4 ml-4">
